Add failing regression test for #5830 main-thread socket-callback freeze
Control-socket command handlers that wait on an async callback (browser
eval/screenshot/cookies WKWebView completions) bridge it to a synchronous
socket reply via `v2AwaitCallback`. On the main thread that helper spun a
nested `CFRunLoopRun()`, freezing the whole app — sidebar plus every other CLI
client serialized behind it — for the full command timeout (the #5830 freeze).

Extract the primitive into a testable file-scope
`socketAwaitCallback(timeout:isMainThread:start:)` (behavior unchanged) and add
a regression test asserting that a main-thread call must return nil immediately
without starting the async work. This commit preserves the buggy nested-runloop
path so the test goes red on CI; the fix lands in the next commit.
