+++ cmuxTests/SocketCallbackAwaiterMainThreadTests.swift
@@ -1,0 +1,56 @@
+import Foundation
+import Testing
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+#elseif canImport(cmux)
+@testable import cmux
+#endif
+
+/// Regression coverage for
+/// https://github.com/manaflow-ai/cmux/issues/5830.
+///
+/// Control-socket command handlers that wait on an async callback (a
+/// `browser eval`/`screenshot`/`cookies` WKWebView completion, etc.) bridge it
+/// to a synchronous socket reply through `socketAwaitCallback`. That waiter
+/// must never run on the **main thread**: historically it spun a nested
+/// `CFRunLoopRun()` there, freezing the whole app (sidebar + every other CLI
+/// client serialized behind it) for the full command timeout.
+///
+/// The control-command execution policy already routes every callback-waiting
+/// command onto the socket-worker thread, so reaching the waiter on the main
+/// thread is a programming error. The contract verified here: a main-thread
+/// call returns `nil` immediately **without** kicking off the async work, so
+/// the dispatcher surfaces a fast timeout instead of parking AppKit.
+@Suite struct SocketCallbackAwaiterMainThreadTests {
+    @Test func mainThreadWaitRefusesToBlockOrStartWork() {
+        nonisolated(unsafe) var startInvoked = false
+        let result: Int? = socketAwaitCallback(timeout: 0.3, isMainThread: true) { _ in
+            // A never-resolving callback. With the old nested-runloop behavior
+            // this branch ran and pinned the thread until the timeout lapsed;
+            // the fix returns before `start` is ever called.
+            startInvoked = true
+        }
+
+        #expect(result == nil)
+        #expect(startInvoked == false)
+    }
+
+    @Test func offMainThreadStillDeliversTheCallbackResult() {
+        // The off-main worker-thread path is unchanged: it must keep blocking on
+        // the callback and returning its value.
+        let result: Int? = socketAwaitCallback(timeout: 1.0, isMainThread: false) { finish in
+            finish(42)
+        }
+
+        #expect(result == 42)
+    }
+
+    @Test func offMainThreadReturnsNilOnTimeout() {
+        let result: Int? = socketAwaitCallback(timeout: 0.05, isMainThread: false) { _ in
+            // Never resolves: the off-main path must time out cleanly to `nil`.
+        }
+
+        #expect(result == nil)
+    }
+}
